Bug fixing and permission utils rework
