+++ app/build.gradle
@@ -146,6 +146,9 @@ dependencies {
     // Date Picker
     implementation 'com.maxkeppeler.sheets-compose-dialogs:core:1.1.1'
     implementation 'com.maxkeppeler.sheets-compose-dialogs:calendar:1.1.1'
+
+    // Permissions
+    implementation "com.google.accompanist:accompanist-permissions:0.30.1"
 }
 
 tasks.withType(Test) {
