+++ app/src/main/java/com/github/geohunt/app/model/database/api/EditedUser.kt
@@ -5,6 +5,7 @@ import android.graphics.Bitmap
 /**
  * Represents a set of modifications on an existing user.
  */
+@Deprecated("Should no longer be used, prefer the repository/view model approach")
 data class EditedUser(
     val user: User,
     var displayName: String,
