+++ app/src/main/java/com/github/geohunt/app/sensor/PermissionUtils.kt
@@ -1,203 +1,89 @@
 package com.github.geohunt.app.sensor
 
-import android.content.Context
-import android.content.pm.PackageManager
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
+import android.Manifest
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material.Button
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalContext
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import com.github.geohunt.app.sensor.PermissionStatus.Denied
-import com.github.geohunt.app.sensor.PermissionStatus.Granted
-import com.github.geohunt.app.utility.findActivity
-import java.util.concurrent.CompletableFuture
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.MultiplePermissionsState
+import com.google.accompanist.permissions.rememberMultiplePermissionsState
 
-/**
- * Defines the two possible status for any permission,
- * can be either [Granted] or [Denied]
- */
-sealed interface PermissionStatus {
-    /**
-     * Defines any permission that has been accepted by the users
-     */
-    object Granted : PermissionStatus
-
-    /**
-     * Defines any permission that has yet to be accepted or been refused by the users
-     */
-    data class Denied(override val shouldShowRationale: Boolean) : PermissionStatus
-
-    /**
-     * Whether we should show a descriptive rationale to explain why the application
-     * needs this specific permission
-     */
-    val shouldShowRationale: Boolean
-        get() = false
-
-    /**
-     * Check whether the current permission is granted
-     */
-
-    val isGranted: Boolean
-        get() = this is Granted
-
-    /**
-     * Check whether the current permission is denied
-     */
-    val isDenied: Boolean
-        get() = this is Denied
-}
-
-/**
- * Define a permission as a 2-Tuple composed of a name and a status
- */
-class Permission(
-    val name: String, val status: PermissionStatus = Denied(false)
-) {
-    /**
-     * Return a permission corresponding with the actual permission state of the provided context
-     */
-    fun updated(context: Context): Permission {
-        val status = when {
-            ContextCompat.checkSelfPermission(context, name) == PackageManager.PERMISSION_GRANTED
-            -> Granted
-
-            ActivityCompat.shouldShowRequestPermissionRationale(context.findActivity(), name)
-            -> Denied(true)
-
-            else
-            -> Denied(false)
-        }
-        return Permission(name, status)
-    }
-
-    /**
-     * Update the permission status with a given value
-     */
-    fun updated(status: PermissionStatus): Permission {
-        return Permission(name, status)
-    }
-}
-
-/**
- * Defines a state representing the state of the permissions that was asked to track
- */
-interface MultiplePermissionState {
-    /**
-     * A list of all the permission that are requested by this state
-     */
-    val permissions: MutableState<List<Permission>>
-
-    /**
-     * Are all permission granted
-     */
-    val allAreGranted: MutableState<Boolean>
-
-    /**
-     * Ask the user for the required permission, once done update the permission list
-     * of this object. The returned future will be completed upon the success of the operation
-     * and will hold an PermissionDeniedException upon failure. Notice that this function
-     * must not be called from the context of a Composable function an should only be called
-     * from a side-effect context
-     */
-    fun requestPermissions(): CompletableFuture<Void>
-}
-
-/**
- * Specialized exception for permission denied handling (using [CompletableFuture])
- */
-class PermissionDeniedException(permissions: List<String>) :
-    RuntimeException("The following permission was denied $permissions")
-
-/**
- * Create a state used to request permission from a composable object. Notice the user must
- * then manually call the launchPermissionRequest on the resulting object
- */
+@OptIn(ExperimentalPermissionsApi::class)
 @Composable
-fun rememberPermissionsState(vararg permissions: String): MultiplePermissionState {
-    val mutableMultiplePermissionState = remember {
-        MutableMultiplePermissionState()
+fun RequirePermissions(
+    permissions: List<String>,
+    withoutPermission: @Composable (MultiplePermissionsState) -> Unit,
+    withPermission: @Composable () -> Unit
+) {
+    val permissionState = rememberMultiplePermissionsState(permissions)
+
+    if (permissionState.allPermissionsGranted) {
+        withPermission()
+    } else {
+        Column {
+
+        }
     }
-    mutableMultiplePermissionState.Initialize(permissions.asList())
-    return mutableMultiplePermissionState
+}
+
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun RequireCameraPermission(withPermission: @Composable () -> Unit) {
+    RequirePermissions(
+        permissions = listOf(Manifest.permission.CAMERA),
+        withoutPermission = { locationPermissionsState ->
+            val textToShow = "GeoHunt requires permission to the camera in order to use this feature"
+            val buttonText = "Request camera permission"
+
+            ShowPermissionRequestPage(locationPermissionsState, textToShow, buttonText)
+        },
+        withPermission
+    )
 }
 
 
-private class MutableMultiplePermissionState() : MultiplePermissionState {
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun RequireFineLocationPermissions(withPermission: @Composable () -> Unit) {
+    RequirePermissions(
+        permissions = listOf(
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.ACCESS_COARSE_LOCATION
+        ),
+        withoutPermission = { locationPermissionsState ->
+            val textToShow = "GeoHunt requires the precise location of this device in order to use this feature"
+            val buttonText = "Request location permission"
 
-    override lateinit var allAreGranted: MutableState<Boolean>
-    override lateinit var permissions: MutableState<List<Permission>>
-    var launcher: ActivityResultLauncher<Array<String>>? = null
-    var future: CompletableFuture<Void>? = null
+            ShowPermissionRequestPage(locationPermissionsState, textToShow, buttonText)
 
-    private val pAreAllGranted : Boolean
-        get() = permissions.value.none { it.status.isDenied }
+        }, withPermission
+    )
+}
 
-    @Composable
-    fun Initialize(perms : List<String>) {
-        val context = LocalContext.current
-
-        permissions = remember {
-            mutableStateOf(perms.map {
-                Permission(it).updated(context)
-            })
-        }
-
-        allAreGranted = remember {
-            mutableStateOf(pAreAllGranted)
-        }
-
-        launcher = rememberLauncherForActivityResult(
-            ActivityResultContracts.RequestMultiplePermissions()
-        ) { result ->
-            updatePermissions(result)
-        }
-    }
-
-    private fun updatePermissions(permissionsUpdated: Map<String, Boolean>) {
-        // First update list of the mutable permission state
-        permissions.value =
-            permissions.value.map {
-                if (permissionsUpdated.getOrElse(it.name) { true }) {
-                    it.updated(Granted)
-                } else {
-                    it.updated(Denied(true))
-                }
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+private fun ShowPermissionRequestPage(
+    locationPermissionsState: MultiplePermissionsState,
+    textToShow: String,
+    buttonText: String
+) {
+    Column {
+        LaunchedEffect(locationPermissionsState) {
+            if (!locationPermissionsState.allPermissionsGranted) {
+                locationPermissionsState.launchMultiplePermissionRequest()
             }
-
-        // Complete the future (either succeed or failed)
-        val deniedPermissions = permissions.value.filter { it.status.isDenied }.map { it.name }
-
-        // Either succeed or deny the future
-        if (deniedPermissions.isEmpty()) {
-            allAreGranted.value = true
-            future?.complete(null)
-        } else {
-            allAreGranted.value = false
-            future?.completeExceptionally(PermissionDeniedException(deniedPermissions))
-        }
-    }
-
-    override fun requestPermissions(): CompletableFuture<Void> {
-        if (future != null) {
-            return future!!
         }
 
-        return if (pAreAllGranted) {
-            CompletableFuture.completedFuture(null)
-        } else {
-            future = CompletableFuture()
-
-            launcher!!.launch(
-                permissions.value.filter { it.status.isDenied }.map { it.name }.toTypedArray()
-            )
-
-            future!!
+        Text(text = textToShow)
+        Spacer(modifier = Modifier.height(8.dp))
+        Button(onClick = { locationPermissionsState.launchMultiplePermissionRequest() }) {
+            Text(buttonText)
         }
     }
 }
